fmt: fix stderr showing up in buffer after writing
Goimport exits with a zero exit value even if there is an error and
output the error to stderr. This was causing the stderr to showing up
in the current user buffer which would mess up the code.

This is fixing the issue by excluding the stderr from the exe command,
see the url below for more info:

http://vim.1045645.n5.nabble.com/how-to-ignore-stderr-in-filter-td1189078.html

This fixes #89 and #147