Add vimproc support for vim-go.
On windows, when call the system() function, a console window will be showed up, which will effect the user exprience terribly.
for example, when the auto-complete was enabled every single time a key pressed a console will interrupt the auto-comple window.
It's a nightmare :(
With the help of vimproc, that nightmare is gone. :)