Initial support for function literals.
Works fine except for the case were a function is returning only a
function (e.g., `func() func() { ... }`). Not sure if there's a good way
around this because that would also be a valid function literal, so the
main issue is trying to sort-of parse this in reverse using just motions
and the character under the cursor doesn't work all that well.

Probably need a forward parser.