tool: fix ShowErrors for various go cmds
We need to cd into the current files directory. This is important so
namemodify does create a full path for outputs when the token is only a
ingle file name (such as for a go test output, i.e.: 'demo_test.go').
For ther outputs, such as 'go install' we already get an absolute path
i.e.: '../foo/foo.go') and fnamemodify successfuly creates the full
ath.)

Fixes #404,#407