go: GOPATH helper commands and improvements for multiple GOPATHs
This PR includes new features and improvements which brings better
support for multiple GOPATHs and tools that support this (such as
godep). This also creates a foundation for other new project based
building tools, such as gb.

The changes are :

* A new :GoPath command that displays the current set GOPATH. It changes
the current GOPATH when an argument is passed, i.e: :GoPath
/home/fatih/mypath
* A new :GoPathClear command that clears the previous set GOPATH via
:GoPath and restores it to the initial GOPATH value. the current set
GOPATH.
* A new go_autodetect_gopath setting. This is enabled
by default. When enabled, vim-go automatically modifies the GOPATH if a
project uses a certain package tool (such as godep). For now only godep
is supported.  If a user edits a file and invokes any vim-go command,
such as :GoDef, :GoBuild, etc.. the GOPATH will include the Godeps
folder. For example any go-to-definition via :GoDef will jump to the
source code inside Godeps.
* Changes to vim-go code base to have better support for multiple OS.