fmt: an experimental in write mode
This uses the `-w` flag, which is available both on gofmt and goimports
to replace the current file instead of outputing it. There is no
rewriting on our side anymore, because gofmt/goimport does it outside
VIM now. This should be also very fast as we don't call anything in Vim.