Do not add log output to quickfix list
Before this change a test that made a call to `log.Println` (or
`Printf`, etc.) resulted in the line printed by `log` to be added to the
quickfix window.

This output:

    === RUN TestSayname
    2015/10/08 13:48:45 log line
    --- FAIL: TestSayname (0.00s)
            testing_example_test.go:9: foobar
    FAIL
    exit status 1
    FAIL    github.com/mrnugget/testing_example     0.005s

would result in these two lines to be added to the quickfix window.

    2015/10/08 13:48:45 log line
            testing_example_test.go:9: foobar

The regex that was responsible for this was meant to only parse the
latter of these two lines, but it accidentally added the first line too.

This commit changes the regex to add lines that have _at least_ one
whitespace at the beginning, thus ignoring the log output.