Add metalinter options
1. Add the option to run the metalinter on save
2. Add the option to limit the output to the currently active buffer

There is a outstanding PR with the `gometalinter` repo to add some
needed logic for this:
https://github.com/alecthomas/gometalinter/pull/71 So I assume this PR
should not be merged before that one is merged.

Also this the first vimscript code I’ve ever written, so I guess this
could be done a little cleaner. Especially the implementation of the
`go#lint#GometaAutoSave` func doesn’t feel all that clean to me, but
I’m not sure how to do this in a different/better way. So if you have