go#package#Paths more efficient at getting GOROOT
go#package#Paths for some reason would call go env GOROOT every single
time it needed the GOROOT. In #697 the problem was that it was calling
go to get the GOROOT every single time the statusline needed an update
which resulted in the flashing cursor. Now it only calls it once and
stores it in a script local variable.