complete: fix ycm
Seems like YCM is setting it's own omnifunc and also started to use
gocode itself recently. This makes vim-go very slugish and also breaks
it. I really hate it as YCM tries to do a lot of things (such as
GoToDefinition). Now it's not compatible anymore with vim-go because it
tries to do the completion itself instead of letting the application do
it.