Make :GoDefPop and :GoDefJump errors better
This is a nice middle ground. :GoDefPop should display a less verbose
message if the user's at the bottom of the stack, as I feel like the
dive-down, pop-out use case will be more common than someone attempting
to jump to an invalid location, and it has a more vimmy feel.

It will still fall back to :GoDefJump's more informative message if a
huge count is given, for example.