Reset shell before running a command in go#util#System (#967)
* Reset shell before running a command in go#util#System

This makes sure the commands still work when using a non-Bourne shell such as
fish or tcsh.

:help vim-doc recommends:

        set shell='/bin/sh'

But this is not a great solution. This will mean that `:sh` will now invoke
`/bin/sh`, instead of `tcsh` or `fish`. It also means that things such as shell
aliases for `:!` will no longer work.