More robust resetting of the shell.
Don't reset it on Windows, and only reset it if `/bin/sh` is executable (which
should pretty much always be the case, but it's not in POSIX and IIRC POSIX
explicitly advises against relying on it).

This is basically the same method that vim-plug uses.

This improves #967 and should fix #986.