coverage: allow to run coverage on _test.go files in subdirs
Due to usage of ':t' modifier to extract filename when switching from
file_test.go to file.go, the original path to file is lost and Vim can't
open it if it's stored in some subdir. It only worked by accident if the
file was located in the same directory.