vim: require at least 8.0.0087 for +job features
+job was introduced in 7.4.x cycle and later at one time it was decided
to release vim with the version number of 8.0. This means that people
having a recent 7.4.x version might have the `+job` feature. But this is
problematic because the initial `+job` implementation is full of bugs.
It was fixed later by many fixes and improvements. 8.0.0087 is the
minimum for *now*, as it fixes another issue about jobs.

We introduce a new wrapper, so we can tweak the requirement ourselves
and tweak the patch level if needed.