handle :GoDef from modified buffer
Handle jumping to type definitions from a modified buffer. When the
definition is in the same file, skip trying to open the file that
contains the definition - just set the cursor position. When the
definition is in a different file and the current buffer has unsaved
modifications, use :hide edit instead of :edit.

Fixes #1125