go_metalinter_deadline applies in async mode now
This change is for issue #1142.

Previous to this change, the option g:go_metalinter_deadline would only
be respected when calling gometalinter if we called it synchronously.

This change removes the default value of g:go_metalinter_deadline of 5
seconds.

For async mode: We pass in a --deadline option now only if we've defined
g:go_metalinter_deadline. For sync mode: We always pass in a --deadline
value. If g:go_metalinter_deadline is not set, we pass in 5 seconds.

Please note that gometalinter itself has a default deadline of 5 seconds
if no deadline is passed to it, so in either case the default of 5
seconds will apply.