Fix aysnchronous :GoReferrers
Fix :GoReferrers so that it populates the location list window correctly
when g:go_auto_sameids = 1. Because the close_cb function was being
referenced using function() instead of funcref(), the callback was
referring to it by name instead of reference; if multiple guru jobs were
executing concurrently, the callback that was executed for all was the
function that was used a closure when the last job was started. Because
the callback function closes over a:args, the value of the last
a:args.custom_parse was used to output the results of all jobs.

Use guru's -modified flag when running asynchronously instead of writing
the current buffer out to a temp file.

Check guru_cmd's return value for a stdin_content key instead of
checking &modified of the current buffer when deciding whether to
pass result.stdin_content in both sync_guru and async_guru.

Remove an unused variable, stdin_content (NOT result.stdin_content!) in
guru_cmd.