use funcref() instead of function() for closures used as callbacks
Use funcref() instead of function() to get a Funcref that will retrieve
the function by reference instead of by name. function() returns a
Funcref that resolves the function by name, which caused multiple
concurrent jobs of the same type to both execute the callback that
closed over the arguments of the last job to be started. funcref()
returns a Funcref that returns the function by reference (not by name),
so each job is assured that its callbacks execute the function that
closed over the arguments of that job.

Also refactored a couple of the guru functions to check the result of
guru_cmd for stdin_content explicitly instead of assuming that
guru_cmd's result has a stdin_content key when the buffer is modified.