Set the filename in the location-list
At somepoint in the near past the location list seems to have broken as
shown by https://github.com/fatih/vim-go/issues/1182. When comparing an
older version of vim-go and master the difference seems to be due to not
including the filename in the location list.

According to the help page for `setqflist()` the filename should be
added to the dictionary as the key `filename`. This lead me to pass the
filename into the function that generates the individual lines of the
location list and then set the file on each error.