job: fix race between channel close and job exit
When channel's close_cb is called, the job this channel is assigned to
may still be alive and the result status of the job will be set to
FAILED (as no actual return code is available at that moment). This is
easily reproduced for me on a single CPU Linux VPS.

To fix this, check for job's status in exit_cb instead.

Signed-off-by: Pavel Borzenkov <pavel.borzenkov@gmail.com>