Load buffers with :GoDef relative to cwd
If I use `gd` on an identifier the filename for the new buffer is set to:

    ~/work/src/github.com/teamwork/desk/server/server.go

Instead of than:

    server/server.go

This makes it open the relative directory, making it behave as `:e
server/server.go`.

I *think* this change should be safe and will work for everyone? I'm not 100% of
that, though. Perhaps we should add a setting?