Show correct message when editing a new file
When editing a new file:

	vim newfile.go

Vim would show the results of the `:0r` command on startup:

	"~/.vim/pack/plugins/start/vim-go/templates/hello_world.go" 7L, 67C

Instead of the default:

	"newfile2.go" [New File]

This corrects that by adding running the `:0r` command with `silent`.

This would only occur if `go#tool#PackageName()` returns `-1`; the correct
message was already shown when a package is found.

There was an `echon` with a comment about removing the `[New File]` message, but
this seems like the wrong behaviour to me. Either way, it didn't work; maybe the
behaviour of `echon` subtly changed?

Fixes #1347