Fix various commands when GOPATH contains spaces.
Test setup:

	export GOPATH="/home/martin/go path/go"
	export PATH="/usr/local/sbin:/usr/local/bin:/usr/bin"

	vim "/home/martin/go path/go/src/a/a.go"

There may be other places as well; this is just the result from a quick test
with my personal default settings. Since there are a lot of if/else branches
relating to jobs etc. it's kind of hard to test everything without a full audit.

It's perhaps a good idea to refactor some of this in a slightly more structured
way at some point: instead of passing strings we should pass a list which is
always escaped (like `os.Exec()`, Python's subprocess, etc.)

This should fix #1288