Fix filepaths in the quickfix list for :GoVet
The `go vet` command would be run with `go#tool#ExecuteInDir()`, which
translates to `cd dir && go vet`.

The problem with this is that if I have `subpkg/file.go` open, the filename that
`go vet` reports will be `file.go`, rather than `subpkg/file.go`. As a result,
the filepath set in the quickfix or location list is also wrong and jumping to
this error will jump to a new (usually non-existing) buffer.

Fix it by explicitly passing the package path, rather than changing the cwd.

Fixes #1376