fix list type fallback behavior
Fix list type fallback behavior so that it correctly handles the
situation where for some command Foo, g:go_list_type's value will be
respected when g:go_list_type_commands is a non-empty dictionary that
does not contain an entry for Foo.

Added a missing command, GoRename, to g:go_list_type_commands.

When displaying the results of a command that was run in a terminal,
pass "_term" to go#list#Type.  _term should not be considered one of the
supported values of g:go_list_type_commands.