Proof read documentation: "Intro", "Install", and README (#1432)
I've been planning on this for a while: proofread all the documentation and
improve where required.

This changes some of the wording to sound better in my opinion. It also adds
various tags where it makes sense, removes or merges some things, and adds a few
minor things.

I'll split it up in several PRs because doing all the documentation in one go is
a lot of work.