refactor for maintainability
Refactor go#list#Type for maintainability. Instead of having the
supported keys (both internally and externally supported) spread
throughout the code, use a single map that contains the keys and their
defaults in autoload/go/list.vim.

Refactor go#list#Type so that it is aware of the map of default values
and does not need its second parameter.