disable g:go_autodetect_gopath by default
Disable g:go_autodetect_gopath by default since first class support for
vendoring was enabled in Go by default a year and a half ago in Go 1.6
and most vendoring tools use it. At this point, there is a greatly
reduced need to detect whether the old method used by Godep and gb is in
effect.