Add go#util#Exec() (#1471)
This is intended as a replacement for `go#util#System()`. The main
difference is that this will accept a list instead of a string as the
command, and that every item in the list will be automatically
shell-escaped.

This means you can just do:

	echo go#util#Exec(['printf', 'look "quotes" and spaces and stuff!'])

and not worry about escaping and stuff.

This approach makes much more sense IMHO, and is also what Go's os/exec
and Python subprocess do. It's especially useful with the jobs API,
since that *doesn't* need escaping, so it makes things easier all round.
Vim still needs to use the shell unfortunately; it can't run programs
directly.

I added a new function instead of modifying `System()` so we can migrate
slowly and carefully; double escaping will break stuff! I just converted
`FillStruct()` as an test/example.

While I'm making a new function anyway I also changed the signature a
bit to something that seems a bit easier to use. I can change that back
if people don't like it or prefer something else.