Clarify the documentation for the guru scope. (#1466)
This clarifies the documentation the guru scope in an FAQ
entry. It also mentions this page when using commands that require the
guru scope.

There is no real way to guess what the user intended, should it be
`github.com/user/pkg`, or `github.com/user/pkg/subpkg`, or something
else?

I think this is a good way to fix #1037.