edit documentation
Edit the INTRO and INSTALL sections of the documentation. Edits mostly
consist of rewording some sentences and paragraphs to read more
naturally.

Consistently capitalize vim-go, even if it's the first word in a
sentence. This might not seem right to you at first, but think about how
other proper nouns that begin with a lower case letter (e.g.  iPhone)
are capitalized when used as the the first word in a sentence.

Make analogous edits to README.md