Don't open the {quickfix,location}list when quitting
In Vim versions before [8.0.1190][1] it would open the new window *and*
leave Vim in the wrong terminal mode (making it unusable).

In later Vim versions it would close the window with the Go file, only
to have a new list popup. Whether this is the desired behaviour is a
matter of opinion, but considering the above bug I think it's better to
*not* open the list for now and simple echo any errors.

This continues from #1497.

[1]: https://github.com/vim/vim/commit/2c33d7bb69c4c2c5b0e39b03cc4b0c04cfdfbb0b