do not try to run tests for :GoTestCompile
Do not try to run a non-existent test for :GoTestCompile. Instead,
compile tests to a temporary file with `go test -c` in order to avoid
problems that may be caused by runtime errors in an init function when
trying to run a non-existent test.

To avoid the problems that 21376e689d6b701b4367cf121da4cfcf17e5fa19
resolved, this change relies on Vim's managing of the temporary
directory instead.