Display warning when people use an older Vim
This shows a warning when people use an older Vim version. You can still set
`g:go_no_version_warning` to override this, if people really want to continue. I
don't think we need to bother documenting that setting in the help page.

The version 7.4.1689 was chosen because that's what the latest Ubuntu LTS
(16.04) uses, which seems like a reasonable choice. I don't mind using an older
version though.