add tests for test output parsing
Add tests for `:GoTest`. While adding tests, it became clear that the
parsing of `go test` errors was mostly correct, but there were a few
edge cases that made writing the tests difficult (namely race conditions
between stdout and stderr when `go test` encountered a panic) that made
the tests flaky. So in addition to adding tests, the parsing of `go
test` output was refactored to use Vim's 'errorformat' option.

Refactor test output parsing to use Vim's 'errorformat' option. Using
'errorformat' makes the parsing of messages more reliable as it keeps
the internal state of whether a multi-line match is being used.
Although the docs for 'errorformat' can be slightly obtuse, they're
actually quite complete and each entry in test.vim's 's:errorformat' is
well-commented now. Using 'errorformat' also provides consistency with
other error parsing used by vim-go (e.g. guru).

Instead of prepending the test name to the test message when
'g:go_test_prepend_name' is set, make the test name its own quickfix
entry, because there's no clear way to sensibly capture the test name
and then use it in the error message. But this is actually quite nice,
because it now includes the test status (e.g. FAIL), and we may be able
to build on this later to also include the passed tests in the error
messages when `:GoTest -v` is used. It also solves the problem of long
test names obscuring the test messages, so it may be more feasible to
set 'g:go_test_prepend_name' by default.

Stack trace produced by panics are now mostly elided in favor of a
single line whose message is the panic message and whose filename and
line number correctly identify the line that produced the panic.