Make :GoDebugStart accept a package name
We need this because otherwise there is no way to debug a non-main
package. For example if `arp242.net/foo` is the main package, and I want
to add a breakpoint to e.g. `arp242.net/foo/bar`.

Also use `g:go_build_tags` and remove `:GoDebugStartWith`; I don't think
we really need to pass arguments to `dlv`, and if we do we can add an
option for it or something.