Fix :GoImpl completion
Fixes #1554

The problem is that `:GoImpl somethingInvalid` would set
`v:shell_error`, and on the next `:GoImpl <Tab>` the `go#util#env()`
call in `s:root_dirs()` would use the cache, so the `v:shell_error` is
"stuck".

Fix it by not checking this here; `go#util#env()` already shows an error
so that should be fine.

Long-term fix is to use `go#util#Exec()` everywhere which returns
`v:shell_error` instead of relying on a global value.

While I'm here also add a basic test and use `Exec()` in a few places.